(리뷰 반영) refactor: SeatJpaRepository 를 별도의 어댑터로 구현체 분리
